+++ custom_debug_derive/src/tests.rs
@@ -14,16 +14,15 @@ fn test_default_struct() {
 
         expands to {
             #[allow(non_upper_case_globals)]
-            #[allow(clippy::match_single_binding)]
             const _DERIVE_core_fmt_Debug_FOR_Point: () = {
                 impl ::core::fmt::Debug for Point {
-                    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+                    fn fmt(&self, fmt: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
                         match self {
                             Point { x: ref __binding_0, y: ref __binding_1, } => {
-                                let mut s = f.debug_struct("Point");
-                                s.field("x", __binding_0);
-                                s.field("y", __binding_1);
-                                s.finish()
+                                let mut debug_builder = fmt.debug_struct("Point");
+                                debug_builder.field("x", __binding_0);
+                                debug_builder.field("y", __binding_1);
+                                debug_builder.finish()
                             }
                         }
                     }
@@ -48,13 +47,13 @@ fn test_format() {
             #[allow(non_upper_case_globals)]
             const _DERIVE_core_fmt_Debug_FOR_Point: () = {
                 impl ::core::fmt::Debug for Point {
-                    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+                    fn fmt(&self, fmt: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
                         match self {
                             Point { x: ref __binding_0, y: ref __binding_1, } => {
-                                let mut s = f.debug_struct("Point");
-                                s.field("x", &format_args!("{:.02}", __binding_0));
-                                s.field("y", __binding_1);
-                                s.finish()
+                                let mut debug_builder = fmt.debug_struct("Point");
+                                debug_builder.field("x", &format_args!("{:.02}", __binding_0));
+                                debug_builder.field("y", __binding_1);
+                                debug_builder.finish()
                             }
                         }
                     }
@@ -81,29 +80,29 @@ fn test_with() {
             #[allow(non_upper_case_globals)]
             const _DERIVE_core_fmt_Debug_FOR_Point: () = {
                 impl ::core::fmt::Debug for Point {
-                    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+                    fn fmt(&self, fmt: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
                         match self {
                             Point { x: ref __binding_0, y: ref __binding_1, } => {
-                                let mut s = f.debug_struct("Point");
-                                s.field("x", &{
+                                let mut debug_builder = fmt.debug_struct("Point");
+                                debug_builder.field("x", {
                                     struct DebugWith<'a, T: 'a> {
                                         data: &'a T,
                                         fmt: fn(&T, &mut ::core::fmt::Formatter) -> ::core::fmt::Result,
                                     }
 
                                     impl<'a, T: 'a> ::core::fmt::Debug for DebugWith<'a, T> {
-                                        fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
-                                            (self.fmt)(self.data, f)
+                                        fn fmt(&self, fmt: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+                                            (self.fmt)(self.data, fmt)
                                         }
                                     }
 
-                                    DebugWith {
+                                    &DebugWith {
                                         data: __binding_0,
                                         fmt: my_fmt,
                                     }
                                 });
-                                s.field("y", __binding_1);
-                                s.finish()
+                                debug_builder.field("y", __binding_1);
+                                debug_builder.finish()
                             }
                         }
                     }
@@ -131,13 +130,13 @@ fn test_skip() {
             #[allow(non_upper_case_globals)]
             const _DERIVE_core_fmt_Debug_FOR_Point: () = {
                 impl ::core::fmt::Debug for Point {
-                    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+                    fn fmt(&self, fmt: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
                         match self {
                             Point { x: ref __binding_0, z: ref __binding_2, .. } => {
-                                let mut s = f.debug_struct("Point");
-                                s.field("x", __binding_0);
-                                s.field("z", __binding_2);
-                                s.finish()
+                                let mut debug_builder = fmt.debug_struct("Point");
+                                debug_builder.field("x", __binding_0);
+                                debug_builder.field("z", __binding_2);
+                                debug_builder.finish()
                             }
                         }
                     }
@@ -163,19 +162,19 @@ fn test_enum() {
             #[allow(non_upper_case_globals)]
             const _DERIVE_core_fmt_Debug_FOR_Foo: () = {
                 impl ::core::fmt::Debug for Foo {
-                    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+                    fn fmt(&self, fmt: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
                         match self {
                             Foo::Bar(ref __binding_0, ref __binding_1,) => {
-                                let mut s = f.debug_tuple("Bar");
-                                s.field(&format_args!("{}i32", __binding_0));
-                                s.field(__binding_1);
-                                s.finish()
+                                let mut debug_builder = fmt.debug_tuple("Bar");
+                                debug_builder.field(&format_args!("{}i32", __binding_0));
+                                debug_builder.field(__binding_1);
+                                debug_builder.finish()
                             }
                             Foo::Quux { x: ref __binding_0, y: ref __binding_1, } => {
-                                let mut s = f.debug_struct("Quux");
-                                s.field("x", __binding_0);
-                                s.field("y", __binding_1);
-                                s.finish()
+                                let mut debug_builder = fmt.debug_struct("Quux");
+                                debug_builder.field("x", __binding_0);
+                                debug_builder.field("y", __binding_1);
+                                debug_builder.finish()
                             }
                         }
                     }
@@ -224,12 +223,12 @@ fn test_bounds_on_skipped() {
                     where
                         TemplatedType<T>: ::core::fmt::Debug
                 {
-                    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+                    fn fmt(&self, fmt: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
                         match self {
                             WantDebug { foo: ref __binding_0, .. } => {
-                                let mut s = f.debug_struct("WantDebug");
-                                s.field("foo", __binding_0);
-                                s.finish()
+                                let mut debug_builder = fmt.debug_struct("WantDebug");
+                                debug_builder.field("foo", __binding_0);
+                                debug_builder.finish()
                             }
                         }
                     }
@@ -270,14 +269,14 @@ fn test_bounds_on_fields_only() {
                         TemplatedType<T>: ::core::fmt::Debug,
                         T: ::core::fmt::Debug
                 {
-                    fn fmt(&self, f: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
+                    fn fmt(&self, fmt: &mut ::core::fmt::Formatter) -> ::core::fmt::Result {
                         match self {
                             WantDebug { foo: ref __binding_0, bar: ref __binding_1, needs_debug: ref __binding_2, } => {
-                                let mut s = f.debug_struct("WantDebug");
-                                s.field("foo", __binding_0);
-                                s.field("bar", __binding_1);
-                                s.field("needs_debug", __binding_2);
-                                s.finish()
+                                let mut debug_builder = fmt.debug_struct("WantDebug");
+                                debug_builder.field("foo", __binding_0);
+                                debug_builder.field("bar", __binding_1);
+                                debug_builder.field("needs_debug", __binding_2);
+                                debug_builder.finish()
                             }
                         }
                     }