+++ CHANGELOG.md
@@ -6,8 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Better error messages
+
 ### Changed
-- potentially breaking: upgrade from edition 2018 to edition 2021
+- BREAKING: upgrade from edition 2018 to edition 2021
 
 ## [0.5.1]